Stop trying
I'm pretty sure you can't do this with griffe alone.
The current python handler's `collect` method
effectively caches the loaded modules.
As a result, if you modify the AST as it's being loaded,
the modification will appear everywhere.
There is no way to only make it appear on one page.
Hence, you have to have a custom handler in addition
to any griffe modifications for this to be able to be
applied on a single page.